+++ Middleware/src/main/java/ca/concordia/encs/citydata/core/contracts/Constants.java
@@ -1,4 +1,4 @@
-package ca.concordia.encs.citydata.core;
+package ca.concordia.encs.citydata.core.contracts;
 
 /***
  * This is the interface for the storing all constants
