+++ Middleware/src/main/java/ca/concordia/encs/citydata/core/contracts/IRunner.java
@@ -1,4 +1,4 @@
-package ca.concordia.encs.citydata.core;
+package ca.concordia.encs.citydata.core.contracts;
 
 /**
  *
