+++ Middleware/src/main/java/ca/concordia/encs/citydata/core/controllers/ApplyController.java
@@ -1,9 +1,10 @@
-package ca.concordia.encs.citydata.core;
+package ca.concordia.encs.citydata.core.controllers;
 
 import java.text.SimpleDateFormat;
 import java.util.Calendar;
 import java.util.Date;
 
+import ca.concordia.encs.citydata.core.contracts.IProducer;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
