+++ Middleware/src/main/java/ca/concordia/encs/citydata/core/controllers/RouteController.java
@@ -1,4 +1,4 @@
-package ca.concordia.encs.citydata.core;
+package ca.concordia.encs.citydata.core.controllers;
 
 import java.util.ArrayList;
 import java.util.List;
