+++ Middleware/src/main/java/ca/concordia/encs/citydata/core/implementations/MiddlewareEntity.java
@@ -1,4 +1,4 @@
-package ca.concordia.encs.citydata.core;
+package ca.concordia.encs.citydata.core.implementations;
 
 import java.util.HashMap;
 import java.util.Set;
