+++ Middleware/src/main/java/ca/concordia/encs/citydata/producers/EnergyConsumptionProducer.java
@@ -2,10 +2,10 @@ package ca.concordia.encs.citydata.producers;
 
 import java.security.InvalidParameterException;
 
-import ca.concordia.encs.citydata.core.AbstractProducer;
-import ca.concordia.encs.citydata.core.IOperation;
-import ca.concordia.encs.citydata.core.IProducer;
-import ca.concordia.encs.citydata.core.IRunner;
+import ca.concordia.encs.citydata.core.implementations.AbstractProducer;
+import ca.concordia.encs.citydata.core.contracts.IOperation;
+import ca.concordia.encs.citydata.core.contracts.IProducer;
+import ca.concordia.encs.citydata.core.contracts.IRunner;
 
 public class EnergyConsumptionProducer extends AbstractProducer<String> implements IProducer<String> {
 	private String city;
