+++ Middleware/src/main/java/ca/concordia/encs/citydata/producers/RandomNumberProducer.java
@@ -3,8 +3,8 @@ package ca.concordia.encs.citydata.producers;
 import java.util.ArrayList;
 import java.util.Random;
 
-import ca.concordia.encs.citydata.core.AbstractProducer;
-import ca.concordia.encs.citydata.core.IProducer;
+import ca.concordia.encs.citydata.core.implementations.AbstractProducer;
+import ca.concordia.encs.citydata.core.contracts.IProducer;
 
 public class RandomNumberProducer extends AbstractProducer<Integer> implements IProducer<Integer> {
 	private int listSize;
